+++ Control_Model_Testing/source/control_model/Control_Model.h
@@ -13,7 +13,7 @@
 #include "..\\state_space_equations\\numerical_ode_solvers.h"
 
 
-int run_sim(double sim_time, double input_time, double dt, int steps,
+int run_sim(int sim_time, double input_time, double dt, int steps,
             double *u, double u_old, double *x_sc, double *x_t,
             double *x_sc_mod, double x_sc_old, double x_t_old,
             double x_sc_mod_old, double *xdot_sc_store, double *xdot_t_store,
